bench/lamicbench_plus/requirements.txt
accelerate
insightface
numpy
onnxruntime-gpu
opencv-python
Pillow
python-dotenv
tabulate
torch
tqdm
transformers>=5.0.0